Reimplement BoM syncing
This got axed when I back-ported the official integration, since it wasn't part of the official feature-set, since this is a thing EMI users actually use a lot, and wasn't a part of any other "item list" mods.

It was an original idea that made sense at the time, and seems like people want it. I am people too, just haven't played much modded minecraft.